requirements.txt
python-dotenv==1.0.0
influxdb-client==1.36.1
psycopg2-binary==2.9.9
plotly==5.15.0
dash==2.14.2
dash-bootstrap-components==1.5.0
scapy==2.5.0
pyinstaller==6.1.0
ping3==4.0.4
pyyaml==6.0.1
python-i18n==0.3.9
psutil==5.9.8
pytest==7.4.3
pytest-cov==4.1.0
black==23.12.1
flake8==7.0.0 